refactor tuple tests into apptest tests where possible
