+++ pypy/objspace/std/test/apptest_specializedtuple.py
@@ -1,0 +1,4 @@
+# spaceconfig = {"objspace.std.withspecialisedtuple": true}
+import sys, os
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+from apptest_tuple import *
